load uniform distribution from file, for result repeatable

+++ qmf/DatasetReader.cpp
@@ -30,8 +30,10 @@ bool DatasetReader::readOne(DatasetElem& elem) {
   if (!std::getline(*stream_, line_)) {
     return false;
   }
-  int32_t value = 0;
-  const int result = sscanf(line_.c_str(), "%ld %ld %d", &elem.userId,
+  double value = 0.0;
+
+  // 输出无所谓，但是对于double scanf必须是%lf
+  const int result = sscanf(line_.c_str(), "%ld %ld %lf", &elem.userId,
       &elem.itemId, &value);
   CHECK_EQ(result, 3) << "the file format is incorrect: " << line_;
   elem.value = static_cast<Double>(value);
